tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
pillow==10.0.1 